pyproject.toml
[project]
name = "hylreg-rag"
version = "0.1.0"
description = "A Retrieval-Augmented Generation system for document question answering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.20,<0.2.0",
    "langchain-community>=0.0.38,<0.1.0",
    "langchain-text-splitters>=0.0.1,<0.1.0",
    "langchain-openai>=0.1.6,<0.2.0",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.0",
    "pypdf>=3.9.1",
    "docx2txt>=0.8",
    "tiktoken>=0.5.1",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.27.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
